refactor: move application core to internal/app and add cmd entrypoint
