+++ .github/workflows/ci.yml
@@ -39,5 +39,5 @@ jobs:
       - name: Build
         run: |
           go build -trimpath \
-            -ldflags "-s -w -X main.version=ci -X main.commit=${GITHUB_SHA::12} -X main.date=$(date -u +%Y-%m-%dT%H:%M:%SZ)" \
-            -o /tmp/opencode2api .
+            -ldflags "-s -w -X github.com/6Kmfi6HP/opencode2api/internal/app.version=ci -X github.com/6Kmfi6HP/opencode2api/internal/app.commit=${GITHUB_SHA::12} -X github.com/6Kmfi6HP/opencode2api/internal/app.date=$(date -u +%Y-%m-%dT%H:%M:%SZ)" \
+            -o /tmp/opencode2api ./cmd/opencode2api
