+++ Makefile
@@ -2,12 +2,12 @@ APP := opencode2api
 VERSION ?= dev
 COMMIT ?= $(shell git rev-parse --short HEAD 2>/dev/null || echo none)
 DATE ?= $(shell date -u +%Y-%m-%dT%H:%M:%SZ)
-LDFLAGS := -s -w -X main.version=$(VERSION) -X main.commit=$(COMMIT) -X main.date=$(DATE)
+LDFLAGS := -s -w -X github.com/6Kmfi6HP/opencode2api/internal/app.version=$(VERSION) -X github.com/6Kmfi6HP/opencode2api/internal/app.commit=$(COMMIT) -X github.com/6Kmfi6HP/opencode2api/internal/app.date=$(DATE)
 
 .PHONY: fmt test vet build release-snapshot clean
 
 fmt:
-	gofmt -w main.go main_test.go
+	gofmt -w ./cmd ./internal
 
 test:
 	go test ./...
@@ -17,7 +17,7 @@ vet:
 
 build:
 	mkdir -p bin
-	go build -trimpath -ldflags "$(LDFLAGS)" -o bin/$(APP) .
+	go build -trimpath -ldflags "$(LDFLAGS)" -o bin/$(APP) ./cmd/opencode2api
 
 release-snapshot:
 	VERSION=$(VERSION) COMMIT=$(COMMIT) DATE=$(DATE) ./scripts/build-release.sh
