+++ docs/responses-compatibility-analysis.md
@@ -7,7 +7,7 @@
 ## 一、数据来源
 
 - 官方规范: `openai-openapi/openapi.yaml` (v3.1.0, ~2.9MB)
-- 代码实现: `main.go`, `responses_protocol.go`, `chat_protocol.go`, `anthropic_protocol.go`
+- 代码实现: `internal/app/main.go`, `internal/app/responses_protocol.go`, `internal/app/chat_protocol.go`, `internal/app/anthropic_protocol.go`
 - 测试覆盖: `request_compatibility_test.go`, `responses_content_test.go`, `responses_builtins_test.go`, `responses_previous_state_test.go`, `responses_store_test.go`, `stream_integrity_test.go`, `protocol_regression_test.go`
 
 ## 二、转换架构
@@ -297,7 +297,7 @@
 
 | 涉及文件 | 说明 |
 |---------|------|
-| `main.go` | `responsesToolFunction` 已有行为：不支持的工具类型返回 false，`convertResponsesTools` 跳过 |
+| `internal/app/main.go` | `responsesToolFunction` 已有行为：不支持的工具类型返回 false，`convertResponsesTools` 跳过 |
 
 ### 9.2 流式 refusal 处理
 
@@ -312,7 +312,7 @@
 
 | 涉及文件 | 修改 |
 |---------|------|
-| `main.go` | `responsesStreamHandler`：新增 refusal 状态、delta 处理、`emitRefusalDone`、`messageItem` 修改 |
+| `internal/app/main.go` | `responsesStreamHandler`：新增 refusal 状态、delta 处理、`emitRefusalDone`、`messageItem` 修改 |
 | `stream_integrity_test.go` | 新增 `TestResponsesStream_RefusalDelta` |
 
 ### 9.3 include 字段功能化
@@ -328,7 +328,7 @@
 
 | 涉及文件 | 修改 |
 |---------|------|
-| `main.go` | `includeHas` 函数、`convertChatToResponses` 签名及逻辑、`responsesHandler` 调用点、`responsesStreamHandler` reasoningItem |
+| `internal/app/main.go` | `includeHas` 函数、`convertChatToResponses` 签名及逻辑、`responsesHandler` 调用点、`responsesStreamHandler` reasoningItem |
 | `request_compatibility_test.go` | 新增 `TestResponsesInclude_EncryptedContentOnlyWhenRequested` |
 | `responses_content_test.go` | 更新 1 处 `convertChatToResponses` 调用 |
 | `protocol_regression_test.go` | 更新 1 处调用 |
@@ -359,8 +359,8 @@
 
 | 涉及文件 | 修改 |
 |---------|------|
-| `main.go` | `ResponsesAPIRequest` 结构体、`ReasonEffort` 结构体、`responsesHandler` extra_body 接线 |
-| `responses_protocol.go` | `applyResponsesRequestEcho` 重写 |
+| `internal/app/main.go` | `ResponsesAPIRequest` 结构体、`ReasonEffort` 结构体、`responsesHandler` extra_body 接线 |
+| `internal/app/responses_protocol.go` | `applyResponsesRequestEcho` 重写 |
 | `request_compatibility_test.go` | 新增 `TestResponsesEcho_MissingFieldsAreEchoed` |
 
 ### 9.5 修复后状态汇总
