+++ internal/app/admin.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"encoding/json"
