+++ internal/app/claude_messages_usage_after_finish_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"encoding/json"
