+++ internal/app/max_tokens_cap_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import "testing"
