+++ internal/app/responses_previous_state_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"encoding/json"
